% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FormatMALDI.R
\name{baseline.sub.ms}
\alias{baseline.sub.ms}
\title{This function uses elements of MALDIquant to baseline subtract from m.z intensity values and create MassSpectrum object from dataframe}
\usage{
baseline.sub.ms(df, name)
}
\arguments{
\item{dataframe}{name}
}
\description{
This function uses elements of MALDIquant to baseline subtract from m.z intensity values and create MassSpectrum object from dataframe
}
\examples{
baseline.sub.ms()
}
\keyword{baseline}

