% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FormatMALDI.R
\name{export.msO}
\alias{export.msO}
\title{This function exports MassSpectrum class objects to a .csv file}
\usage{
export.msO(msobject, file)
}
\arguments{
\item{MassSpectrum-class}{file}
}
\description{
This function exports MassSpectrum class objects to a .csv file
}
\examples{
export.msO()
}
\keyword{export}

