% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FormatMALDI.R
\name{format.spectra.df}
\alias{format.spectra.df}
\title{This function formats spectra data so that dataframe contains only m/z time and intensity}
\usage{
\method{format}{spectra.df}(data.frame, sample, feature)
}
\arguments{
\item{dataframe,}{sample, feature}
}
\description{
This function formats spectra data so that dataframe contains only m/z time and intensity
}
\examples{
format.spectra.df()
}
\keyword{format}
\keyword{spectra}

