% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FormatMALDI.R
\name{importWorksheets}
\alias{importWorksheets}
\title{Importing Excel Document Function}
\usage{
importWorksheets(file)
}
\arguments{
\item{.xls}{file}
}
\description{
This Function converts multi sheet excel doc to single data-frame with m/z,intensity,
}
\examples{
importWorksheets()
}
\keyword{import}

