% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FormatMALDI.R
\name{Plot.spectra.intens}
\alias{Plot.spectra.intens}
\title{This function plots intensity by m.z values}
\usage{
Plot.spectra.intens(df = data.frame)
}
\arguments{
\item{dataframe}{}
}
\description{
This function plots intensity by m.z values
}
\examples{
Plot.spectra.intens()
}
\keyword{plot}

