% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FormatMALDI.R
\name{plot.spectra.rel.intens}
\alias{plot.spectra.rel.intens}
\title{This function plots relative intensity by m.z values}
\usage{
plot.spectra.rel.intens(df = data.frame)
}
\arguments{
\item{dataframe}{}
}
\description{
This function plots relative intensity by m.z values
}
\examples{
plot.spectra.rel.intens()
}
\keyword{plot}

