% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FormatMALDI.R
\name{Return.spectra.list}
\alias{Return.spectra.list}
\title{This function converts all of the sheets from an excel file as a single list and returns a list called spectra.list containing all sheets in excel file to global enironment it also returns sheet names so user can create dataframes containing peak lists of interest}
\usage{
Return.spectra.list(file)
}
\arguments{
\item{file}{}
}
\description{
This function converts all of the sheets from an excel file as a single list and returns a list called spectra.list containing all sheets in excel file to global enironment it also returns sheet names so user can create dataframes containing peak lists of interest
}
\examples{
Return.spectra.list()
}
\keyword{spectra}

